+++ tests/unit/utils/build-popup-content-test.js
@@ -1,0 +1,10 @@
+import buildPopupContent from 'labs-regional-viz/utils/build-popup-content';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | build popup content');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = buildPopupContent();
+  assert.ok(result);
+});
